Allow waiting single interface
Some systems may have network interfaces which never get up. This would
prevent the process from starting. This patch allows to specify a single
interface to wait for.
If NFS_WAITING_INTERFACE is specified, this is the only interface the
process will wait for before trying to mount NFS drive. If the variable
is kept empty (default value), then the default behavior will be the
same as today, i.e. waiting for all interfaces to be ready.